refactor some code

+++ src/app/details-fiche/details-fiche.component.ts
@@ -1,74 +1,83 @@
-import { UnsubscribeBase } from './../shared/unsubscribeBase';
-import { takeUntil } from 'rxjs/operators';
-import { ConfirmationService, MessageService } from 'primeng/api';
+import { GenericCrud } from './../shared/generic-crud';
+import { takeUntil, switchMap } from 'rxjs/operators';
+import { ConfirmationService } from 'primeng/api';
 import { Location } from '@angular/common';
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { FichesService } from '../services/fiches.service';
 import { map } from 'rxjs';
 import { Fiche, Formules } from '../types';
+import { FormBuilder } from '@angular/forms';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-details-fiche',
   templateUrl: './details-fiche.component.html',
   styleUrls: ['./details-fiche.component.scss'],
 })
-export class DetailsFicheComponent extends UnsubscribeBase<Fiche> {
-  ficheId: string | null;
-  ficheDetails: Fiche | undefined;
+export class DetailsFicheComponent
+  extends GenericCrud<Fiche>
+  implements OnInit
+{
+  ficheId: string | null | undefined;
+  klantName: string | undefined;
+  KlantDetails: Fiche | undefined;
   formulesList: Formules[] | undefined;
   constructor(
-    private ficheService: FichesService,
-    public route: ActivatedRoute,
-    private location: Location,
-    private messageService: MessageService,
-    private confirmationService: ConfirmationService,
-    private router: Router
+    protected override ficheService: FichesService,
+    protected override route: ActivatedRoute,
+    protected override confirmationService: ConfirmationService,
+    protected override router: Router,
+    protected override fb: FormBuilder,
+    protected override activeRoute: ActivatedRoute,
+    private toastr: ToastrService
   ) {
-    super();
+    super(ficheService, route, confirmationService, router, fb, activeRoute);
+  }
+
+  ngOnInit(): void {
     this.ficheId = this.route.snapshot.paramMap.get('id');
-    this.getFicheById(String(this.ficheId));
-    this.getAllFormulesOnFicheId(String(this.ficheId));
+    this.getKlantDetailsById();
+    this.getAllFormulesOnFicheId();
   }
 
   // delete a fiche
   deleteFiche(): void {
     if (this.ficheId) {
       this.confirmationService.confirm({
-        message: 'Weet je zeker dat je dit klantenfiche wil verwijderen?',
+        message: `Weet je zeker dat je het fiche van ${this.klantName} wil verwijderen?`,
         accept: () => {
           this.ficheService.deleteFiche(String(this.ficheId)).then(() => {
-            this.location.back();
+            this.router.navigate(['/fiches']).then(() => {
+              this.toastr.info('Het fiche is verwijderd', 'Verwijderd');
+            });
           });
         },
       });
     }
   }
 
-  // update a fiche
-  updateFiche(): void {
-    this.router.navigate(['fiches/new', this.ficheId]);
-  }
-
   // get a fiche on id
-  getFicheById(id: string): void {
+  getKlantDetailsById(): void {
     if (this.ficheId) {
       this.ficheService
-        .getFicheById(id)
+        .getKlantDetailsById(this.ficheId)
         .pipe(
-          map((data) => data.payload.data()),
+          switchMap(async (data) => data.payload.data()),
           takeUntil(this.destroy$$)
         )
         .subscribe((data) => {
-          this.ficheDetails = data;
+          this.klantName = data?.achternaam + ' ' + data?.voornaam;
+          this.KlantDetails = data;
         });
     }
   }
-  formuleId: string | undefined;
-  getAllFormulesOnFicheId(id: string): void {
+
+  // get all formules on fiche id
+  getAllFormulesOnFicheId(): void {
     if (this.ficheId) {
       this.ficheService
-        .getAllFormulesOnFicheId(id)
+        .getAllFormulesOnFicheId(this.ficheId)
         .snapshotChanges()
         .pipe(
           map((changes) =>
@@ -84,10 +93,17 @@ export class DetailsFicheComponent extends UnsubscribeBase<Fiche> {
         });
     }
   }
+
+  // remove a formule on id
   removeFormules(item: Formules): void {
     if (item.id) {
-      this.ficheService.removeFormules(item.id).then(() => {
-        console.log('deleted');
+      this.confirmationService.confirm({
+        message: `Weet je zeker dat je het formule wil verwijderen? Dit is definitief en kan later niet opnieuw opgevraagd worden!`,
+        accept: () => {
+          this.ficheService.removeFormules(item.id).then(() => {
+            this.toastr.info('De formule is verwijderd', 'Verwijderd');
+          });
+        },
       });
     }
   }
